GETP-216 refactor: application serivce가 in-port를 구현하도록 변경
